server/requirements.txt
fastapi
uvicorn
pydantic
pyinstaller
python-escpos
pyusb
pillow
